requirements.txt
asgiref==3.2.5
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.3
djangorestframework==3.12.1
drf-yasg==1.20.0
Flask==1.1.1
idna==2.9
inflection==0.5.1
itsdangerous==1.1.0
itypes==1.2.0
Jinja2==2.11.1
MarkupSafe==1.1.1
packaging==20.4
pyparsing==2.4.7
PyQt5==5.14.1
PyQt5-sip==12.7.1
pytz==2019.3
requests==2.23.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.3.0
uritemplate==3.0.1
urllib3==1.25.8
Werkzeug==1.0.0
